Invert loop
Small refactor to more easily allow us to provide an `update_resource` action that can chain with other transforms.

This passes the existing tests but I need to add some more coverage before I can be confident that this is a non-breaking change.

One caveat with this is that I'm buffering the whole slice of ResourceMetrics up front. Not sure of the performance implications.